Trying to fix segv

+++ src/common_cpp/Recon/SciFi/RealDataDigitization.cc
@@ -65,15 +65,15 @@ void RealDataDigitization::process(Spill *spill, Json::Value const &daq) {
     if ( tracker_event[i].isMember("VLSB_C") ) {
       Json::Value input_event = tracker_event[i]["VLSB_C"];
       // Merge tracker events.
-      for ( size_t idig = 0; idig < daq["tracker2"][i]["VLSB_C"].size(); ++idig ) {
-        input_event[input_event.size()] = daq["tracker2"][i]["VLSB_C"][idig];
+      for ( int idig = 0; idig < int(daq["tracker2"][i]["VLSB_C"].size()); ++idig ) {
+        input_event[int(input_event.size())] = daq["tracker2"][i]["VLSB_C"][idig];
       }
       process_VLSB_c(input_event, event, tracker0daq_event, tracker1daq_event);
     } else if ( tracker_event[i].isMember("VLSB") ) {
       Json::Value input_event = tracker_event[i]["VLSB"];
       // Merge tracker events.
-      for ( size_t idig = 0; idig < daq["tracker2"][i]["VLSB"].size(); ++idig ) {
-        input_event[input_event.size()] = daq["tracker2"][i]["VLSB"][idig];
+      for ( int idig = 0; idig < int(daq["tracker2"][i]["VLSB"].size()); ++idig ) {
+        input_event[int(input_event.size())] = daq["tracker2"][i]["VLSB"][idig];
       }
       process_VLSB(input_event, event, tracker0daq_event, tracker1daq_event);
     } else {
